rusk: build.rs to use dusk-plonk 0.7
The new verifier data structure of dusk-plonk 0.7 leads to major
optimizations in the proof verification, but the build strategy of the
keys must change.

Besides that, the previous version of the build script was not
benefiting from release optimizations and this can lead to huge time
loss considering the circuts compilation process is CPU intensive.

Resolves #261