transfer-contract: split `execute` in two stages
To allow for splitting execution of a transaction, we split the
`execute` function into two different functions - `spend` and `refund`.

The `spend` function checks that a transaction can spend its inputs and
produce its outputs, and execute the transfer, while the `refund`
function refunds a specific amount to the previously performed
transaction.

This allows the host to execute code in between the two calls, such as
executing a contract call, and decide exactly what a transaction should
be charged based on the results.

See also: #939