contracts: change 'hard slash' mechanism
Change the code to slash the stake incrementally, depending on
the number of 'hard_faults'. The stake is slashed by 10% times
the number of faults. The 'severity' parameter can be used to
increase 'hard_faults' by more than 1.

The slashed stake is also suspended for the resto of the current
epoch plus 'hard_faults'additional epochs.

See also #1651