transfer-contract: support for moonlight and expanded withdrawals
We add support for Moonlight transactions by adding a map between public
keys and their respective account data as defined by `execution-core`.
This means also adding a query function for this account data - named
`account`, taking as argument a public key.

Support for the new withdrawal system is also added, in a way that we
can reuse code between `withdraw` and `mint`, effectively minimizing the
amount of code written. One downside of this new system is that
contracts are no longer able to set data in the `Sender` of a Phoenix
note. Instead the tranfer contract is responsible for setting it, and
currently we set the contract ID withdrawn from, together with a
suffixed string.

We take the liberty to rework the transient data held during a call ray,
effectively minimizing data being passed back and forth - for instance
during a `refund` call. We also rename `balance` to `contract_balance`
contract-wide, to avoid confusion with Moonlight accounts.